task 4

+++ bit_manipulation/4-clear_bit.c
@@ -10,7 +10,7 @@ int clear_bit(unsigned long int *n, unsigned int index)
 	if (index > 63)
 		return (-1);
 
-	*n &= ~((unsigned long int )1 << index);
+	*n &= ~(1 << index);
 
 		return (1);
 }
